tweaks to improve ease of dev and included express in the package.json

+++ start-dev.sh
@@ -1,4 +1,8 @@
 #!/bin/sh
 
-docker build -t bolts-site:local .
-docker run -e GTAG=UA-131204488-2 -p 8000:80 --name bolts-site --rm bolts-site:local
+if [ -d "node_modules" ]; then
+    docker build -t bolts-site:local .
+    docker run -e GTAG=UA-131204488-2 -p 8000:80 --name bolts-site --rm bolts-site:local
+else
+    echo "Error: Please run \"npm install\" before executing this script"
+fi
